Eliminate the storage of the bridging trin nodes
There is no need to store anything, since their purpose is to offer
content to others. This eliminates local storage used during bridging.

Most importantly, it means that these nodes won't indicate interest in
content to each other. We want to know that we have pushed content
elsewhere in the network, after offering.