Update CLI parsing, to prepare for block range arg
We want more explicit commands given to the bridge module, so that it's
clear whether we are following the latest on the chain, or injecting a
file, or backfilling an old block number.

Update docs with the new CLI argument structure

Add release note mentioning the breaking change to --latest, for
following the head of the chain.